requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
celery>=5.3.0
drf-yasg>=1.21.7
django-environ>=0.11.0
mysqlclient>=2.2.0

